Extracted AppManagement from NetDaemonHost
